+++ apps/test/src/config/config.common.ts
@@ -18,6 +18,7 @@ export const appConfigCommon: AppConfigCommon = {
     iam: {
       accessTokenExpiryTimeInMinutes: 120,
       defaultUserIdentifierField: 'id',
+      oauth2: {},
       userPasswordHMACAlgorithm: 'sha256',
       refreshTokenExpiryTimeInHours: 24
     }
@@ -28,7 +29,7 @@ export const appConfigCommon: AppConfigCommon = {
     cache: {
       defaultTTL: 3600,
       defaultIndividualSearchEnabled: true,
-      storeKey: Constants.PERSISTANCE_CACHE_STORE_KEY,
+      storeKey: Constants.DATA_CACHE_STORE_KEY,
       ttlPerEntity: { users: 60000 },
       type: NoSQLType.Redis,
       useHashmap: false
@@ -40,17 +41,17 @@ export const appConfigCommon: AppConfigCommon = {
       // failOnConnectionError: false,
       // sentinelMode: true,
       storeDelimiter: ':',
-      storeKey: Constants.PERSISTANCE_CACHE_AUTH_STORE_KEY,
+      storeKey: Constants.DATA_CACHE_AUTH_STORE_KEY,
       type: NoSQLType.Valkey,
       useHashmap: true
     },
     db: {
-      connectionName: Constants.PERSISTANCE_DB_MODULE_CONNECTION_NAME,
+      connectionName: Constants.DATA_DB_MODULE_CONNECTION_NAME,
       // failOnConnectionError: false,
       type: RDBType.MySQL
     },
     dbConfigs: {
-      connectionName: Constants.PERSISTANCE_DB_CONFIGS_MODULE_CONNECTION_NAME,
+      connectionName: Constants.DATA_DB_CONFIGS_MODULE_CONNECTION_NAME,
       // failOnConnectionError: false,
       type: RDBType.MySQL
     }