+++ packages/core/src/common/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './base64UrlEncode';
 export * from './getNested';
 export * from './loadDynamicModules';
 export * from './setNested';