+++ packages/domain-iam/src/common/definitions/common.constants.ts
@@ -1,4 +1,8 @@
 export enum Constants {
   // eslint-disable-next-line no-unused-vars
-  DOMAIN_MODULE_NAME = 'DOMAIN_MODULE_NAME'
+  DOMAIN_MODULE_NAME = 'DOMAIN_MODULE_NAME',
+  // eslint-disable-next-line no-unused-vars
+  OAUTH2_CODE_VERIFIER_LENGTH = 128,
+  // eslint-disable-next-line no-unused-vars
+  OAUTH2_PKCE_CHALLENGE_HASH_METHOD = 'SHA-256'
 }