+++ packages/domain-iam/src/services/authenticationOAuth2/iam.authenticationOAuth2.definitions.ts
@@ -12,7 +12,13 @@ export interface OAuth2AuthenticateUserResult extends AuthenticateUserResult {
   accessCode: string;
 }
 
-export interface OAuth2GenerateAuthorizationURL {
+export interface OAuth2GenerateAuthorizationURLData {
+  scope?: string;
+}
+
+export interface OAuth2GenerateAuthorizationURLReturnData {
   authorizationCodeRequestURL: string;
   codeChallenge: string;
+  codeVerifier: string;
+  state: string;
 }