+++ packages/domain-iam/src/services/index.ts
@@ -1,5 +1,6 @@
 export * from './authentication';
 export * from './authenticationLocal';
 export * from './authorization';
+export * from './oAuth2Provider';
 export * from './tokenManager';
 export * from './users';