+++ packages/domain-iam/src/services/oAuth2Provider/iam.oAuth2Provider.definitions.ts
@@ -1,0 +1,5 @@
+export interface GetAuthorizationUrlData {
+  challenge: string;
+  scope?: string;
+  state: string;
+}