+++ packages/domain-iam/src/services/oAuth2Provider/index.ts
@@ -1,0 +1,2 @@
+export * from './iam.oAuth2Provider.definitions';
+export * from './iam.oAuth2Provider.service';